Refactor Generative component: move content to GenerativeContent for better structure
