+++ DashAI/front/src/pages/generative/Generative.jsx
@@ -1,133 +1,8 @@
-import { Box } from "@mui/material";
-import { useState, useEffect } from "react";
-import ModuleContainer from "../../components/layout/ModuleContainer";
-import LeftPanel from "../../components/threeSectionLayout/panels/LeftPanel";
-import CenterPanel from "../../components/threeSectionLayout/panels/CenterPanel";
-import RightPanel from "../../components/threeSectionLayout/panels/RightPanel";
-import SessionBar from "../../components/generative/SessionBar";
-import SelectTaskMenu from "../../components/generative/SelectTaskMenu";
-import GenerativeChat from "../../components/generative/GenerativeChat";
-import SelectModelMenu from "../../components/generative/SelectModelMenu";
-import ParamsBar from "../../components/generative/ParamsBar";
-import { getSessions, removeSession } from "../../api/session";
-import { getComponents } from "../../api/component";
-import { useTranslation } from "react-i18next";
-import { useThreePanelLayout } from "../../hooks/useThreePanelsLayout";
-import { ThreePanelLayoutContext } from "../../components/threeSectionLayout/panels/ThreePanelLayoutContext";
+//import { TourProvider } from "../../components/tour/TourProvider";
+//import { TOUR_KEYS } from "../../constants/tours";
+//import { DatasetsAndNotebooksProvider } from "../../components/custom/contexts/DatasetsAndNotebooksContext";
+import GenerativeContent from "./GenerativeContent";
 
 export default function Generative() {
-  const [stepIndex, setStepIndex] = useState(0);
-  const [selectedSessionId, setSelectedSessionId] = useState(null);
-  const [tasks, setTasks] = useState([]);
-  // TODO: Combine selectedTaskName and selectedDisplayName into a single selectedTask State
-  const [selectedTaskName, setSelectedTaskName] = useState("");
-  const [selectedDisplayName, setSelectedDisplayName] = useState("");
-  const [sessions, setSessions] = useState([]);
-  const [paramsVersion, setParamsVersion] = useState(0);
-  const { t } = useTranslation(["generative", "common"]);
-  const threePanelLayout = useThreePanelLayout();
-
-  const handleSessionClick = (sessionId, taskName, taskDisplayName) => {
-    setSelectedTaskName(taskName);
-    setSelectedSessionId(sessionId);
-    setSelectedDisplayName(taskDisplayName);
-  };
-
-  const handleNewSessionButton = () => {
-    setSelectedSessionId(null);
-    setStepIndex(0);
-    setSelectedTaskName("");
-  };
-
-  const onParamsUpdate = (newParams) => {
-    setParamsVersion((prev) => prev + 1);
-  };
-
-  useEffect(() => {
-    getSessions().then((data) => {
-      setSessions(data);
-    });
-  }, []);
-
-  useEffect(() => {
-    try {
-      getComponents({ selectTypes: "GenerativeTask" }).then((data) => {
-        setTasks(data);
-      });
-    } catch (error) {
-      console.error("Error fetching generative tasks:", error);
-    }
-  }, [t]);
-
-  const handleAddSession = (session) => {
-    setSessions((prevSessions) => [session, ...prevSessions]);
-  };
-
-  const handleSessionDelete = (id) => {
-    if (id === selectedSessionId) {
-      setSelectedSessionId(null);
-      setStepIndex(0);
-    }
-
-    setSessions((prevSessions) =>
-      prevSessions.filter((session) => session.id !== id),
-    );
-
-    removeSession(id);
-  };
-
-  return (
-    <ThreePanelLayoutContext.Provider value={threePanelLayout}>
-      <ModuleContainer>
-        <LeftPanel>
-          <SessionBar
-            tasks={tasks}
-            sessions={sessions}
-            selectedSessionId={selectedSessionId}
-            handleSessionClick={handleSessionClick}
-            handleNewSessionButton={handleNewSessionButton}
-            handleSessionDelete={handleSessionDelete}
-            stepIndex={stepIndex}
-            onToggle={threePanelLayout.handleToggleLeft}
-          />
-        </LeftPanel>
-        <CenterPanel>
-          {selectedSessionId ? (
-            <GenerativeChat
-              sessionId={selectedSessionId}
-              taskName={selectedTaskName}
-              paramsVersion={paramsVersion}
-            />
-          ) : stepIndex === 0 ? (
-            <SelectTaskMenu
-              tasks={tasks}
-              goToNextStep={(taskName, displayName) => {
-                setSelectedDisplayName(displayName);
-                setSelectedTaskName(taskName);
-                setStepIndex(1);
-              }}
-            />
-          ) : (
-            <SelectModelMenu
-              goToBackStep={() => setStepIndex(0)}
-              handleAddSession={handleAddSession}
-              selectedTaskName={selectedTaskName}
-              selectedDisplayName={selectedDisplayName}
-              setSelectedSessionId={setSelectedSessionId}
-              existingSessions={sessions}
-            />
-          )}
-        </CenterPanel>
-
-        <RightPanel toggleButtonTop="50%">
-          <ParamsBar
-            selectedSessionId={selectedSessionId}
-            onParamsUpdate={onParamsUpdate}
-            taskName={selectedTaskName}
-            onToggle={threePanelLayout.handleToggleRight}
-          />
-        </RightPanel>
-      </ModuleContainer>
-    </ThreePanelLayoutContext.Provider>
-  );
+  return <GenerativeContent />;
 }
